+++ src/pytorch_lightning/CHANGELOG.md
@@ -139,7 +139,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - The `WandbLogger` will now use the run name in the logs folder if it is provided, and otherwise the project name  ([#12604](https://github.com/PyTorchLightning/pytorch-lightning/pull/12604))
 
 
--
+- The `LightningModule.save_hyperparameters` method now asserts that it is called properly. Calls from outside the `LightningModule.__init__` method will result in a `RuntimeError` to protect checkpointing functionality.
 
 
 ### Deprecated